# src/entity/user.entity.ts
import { Entity, PrimaryGeneratedColumn, Column } from 'typeorm'
import { Base } from './base.entity.js'

@Entity()
export class User extends Base {
    @PrimaryGeneratedColumn()
    id: number

    @Column({ type: "varchar", length: 50, nullable: false })
    name: string

    @Column({ type: "varchar", length: 100, nullable: false, unique: true })
    email: string

    @Column({ type: "varchar", length: 100, nullable: false })
    password: string
}